change parsing logic and fix field, tested solution through postman

+++ src/main/kotlin/com/example/auditapi/model/AuditIssueDto.kt
@@ -5,7 +5,7 @@ import java.time.OffsetDateTime
 
 data class AuditIssueDto(
     val updateDate: String,
-    val line: Int,
+    val line: Double,
     val rule: String,
     val project: String,
     val effort: String,
